changed the things according to the suggestions
