+++ src/assets/global.css
@@ -1,4 +1,4 @@
-@import url("https://fonts.googleapis.com/css2?family=Open+Sans:wght@300;400;600;700&display=swap");
+
 
 html {
   font-family: Arial, Helvetica, sans-serif;
@@ -20,5 +20,50 @@ body {
   position: absolute;
   right: 0;
   top: 0;
+}
+.button {
+  vertical-align: middle;
+  display: inline-block;
+  position: absolute;
+  border-radius: 2%;
+  background-color: white;
+  border: rgb(0, 0, 0);
+  color: #000000;
+  text-align: center;
+  font-size: 28px;
+  padding: 20px;
+  width: 300px;
+  transition: all 0.5s;
+  cursor: pointer;
+  margin: 5px;
+  margin-left: 42%;
+  margin-top: 20%;
+  text-align: center;
+  z-index: 999;
+}
+
+.button span {
+  cursor: pointer;
+  display: inline-block;
+  position: relative;
+  transition: 0.5s;
+}
+
+.button span:after {
+  content: "\00bb";
+  position: absolute;
+  opacity: 0;
+  top: 0;
+  right: -20px;
+  transition: 0.5s;
+}
+
+.button:hover span {
+  padding-right: 25px;
+}
+
+.button:hover span:after {
+  opacity: 1;
+  right: 0;
 }
